+++ src/components/ids-menu-button/ids-menu-button.ts
@@ -48,7 +48,7 @@ export default class IdsMenuButton extends Base {
     if (this.hasAttribute(attributes.DROPDOWN_ICON)) {
       this.#configureDropdownIcon(true);
     }
-    this.configureMenu();
+    this.#initMenuPopup();
   }
 
   disconnectedCallback() {
@@ -190,6 +190,16 @@ export default class IdsMenuButton extends Base {
     let el = findPopup(this.parentElement);
     if (!el) el = findPopup(this.getRootNode());
     return el;
+  }
+
+  /**
+   * Set popup menu initially
+   * @private
+   * @returns {void}
+   */
+  #initMenuPopup(): void {
+    this.configureMenu();
+    if (!this.menuEl?.popup) requestAnimationFrame(() => this.configureMenu());
   }
 
   /**